Fix PR#282 review: content hash DB with bloom filter
- Fix check_and_add() bug: verify with _hash_set for O(1) lookup before bloom
- Add flush() method for batch saving (saves to disk only when needed)
- Decouple: move BloomFilter and ContentHashDB to dedicated content_hash.py
- Update tests to use pytest tmp_path fixture
- Fix exception handling to (OSError, IOError)
- Add file_already_mined() fallback to ChromaDB as safety net
